ai-backend-ergo-app-prod/requirements.txt
django==5.0
djangorestframework==3.14.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.8
tensorflow==2.15.0
numpy==1.26.3
gunicorn==21.2.0
torch==2.1.2
reportlab==4.0.8
transformers==4.36.2
django-cors-headers==4.3.1
opencv-python==4.8.1.78
channels==4.0.0
anthropic>=0.18.1
python-dotenv>=1.0.0
ultralytics==8.0.227
pandas>=2.1.0  # Added for YOLOv5 results processing